Removed category from BaseDefinitionBuilder Move helpers to GuiPresentationBuilder
